Got the OCC led working, it needs its own direct coupling to the CMP output, looks like compiler optimization when putting it in processio(). Occupied detection is working, on the proto the blue led goes on.
Auto updated of amp data is still only done when sending new PWM commands, so the added timer to push with 2Hz back is not updating the UI...

+++ TrackAmplifier4.X/modbus/PetitModbus.h
@@ -33,7 +33,7 @@
 extern unsigned char PETITMODBUS_SLAVE_ADDRESS;                                 // Petit Modbus RTU Slave icin adres numarasi [0 to 255]
 
 typedef struct{
-            unsigned int                     ActValue;
+            uint16_t                     ActValue;
         }PetitRegStructure;
 
 #if ((PETITMODBUS_READ_HOLDING_REGISTERS_ENABLED > 0)|| (PETITMODBUSWRITE_SINGLE_REGISTER_ENABLED > 0) || (PETITMODBUS_WRITE_MULTIPLE_REGISTERS_ENABLED > 0))
